requirements.txt
fastapi
uvicorn
pandas
scikit-learn
openai
numpy
supabase
python-dotenv
